% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wrds.R
\name{applyFilters}
\alias{applyFilters}
\title{Apply filters to WRDS query}
\usage{
applyFilters(...)
}
\arguments{
\item{...}{A list of filters generated using the filter_* functions}
}
\description{
A helper function to easily specify one or several filters. All specified filters are joined by "and" (i.e., all must be true).
}
