% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wrds.R
\name{filter_lt}
\alias{filter_lt}
\title{Filter less than}
\usage{
filter_lt(x, condition)
}
\arguments{
\item{x}{The name of the variable on which the condition should be evaluated}

\item{condition}{The value to which the filter should compare}
}
\description{
All filter_* functions restrict your query based on the value of a variable x, as compared to a condition. The second part of the functions' names (after the underscore) match the comparison types available in the WRDS API. For more info, read: https://wrds-api.wharton.upenn.edu/tutorial-data-operators/
}
