% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wrds.R
\name{getTokenFromHome}
\alias{getTokenFromHome}
\title{Get WRDS API key from home directory}
\usage{
getTokenFromHome()
}
\description{
Looks for a .wrdstoken file in the user's home directory and, if present, returns it. If not, an error is thrown. Note: to function properly, the .wrdstoken file must contain the user's API key ONLY. This key can be found at wrds-api.wharton.upenn.edu.
}
