% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getURLvector}
\alias{getURLvector}
\title{Returns WRDS url vector based on parameters}
\usage{
getURLvector(identifier, filters, root, count, limit = 2000)
}
\description{
Returns WRDS url vector based on parameters
}
\keyword{internal}
